requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-matplotlib==0.1
idna==2.9
networkx==2.4
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
plotly==4.6.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
